requirements.txt
Flask==2.3.2
asgiref==3.7.2
requests==2.31.0
PyJWT==2.7.0
uvicorn==0.22.0